I spent a few hours working how the rooms work in GMS2. I was going to have a 2d array storing everything, but I can take a few shortcuts and save myself some time and trouble. I'm working on changing my napkin notes over to using direct spritebox collisions instead of digging through lists of stuff (mostly because I can't figure out how to make grids of lists in GMS2, but that'll be for another day). For now, I've got a decent handle on how to make collisions. I didn't do the break turns, so those are still on the list.
TO DO MK 2:
 - Break Turns
 - basic gui for messages and mouse-overs
 - mouse over tooltips
 - mouse movement (this is gonna be bad)
 - Figure out how I'm gonna do randomization

First major goal will be to have a level that teaches you how to fly, since it's not very intuitive at the start. Tutorial missions are they way to go imo, means I don't need to do a bunch
of random-gen and can work on the important part of 'does thing work'.